Support atomic chess without checks (#81)
Add variant nocheckatomic which implements a modified version
of atomic chess where check and check-/stalemate do not apply,
which is very similar to the rules used on ICC.

The variant atomic is an incomplete implementation of
lichess atomic chess rules, added for easier usage in GUIs.
It will occasionally play illegal moves in this variant.