Generalize NNUE architecture
Remove restrictions on board size and piece types
in order to enable NNUE for most supported variants.

This change modifies the indexing for fairy pieces,
therefore existing networks for fairy variants, currently only Makruk,
will not work properly any more and need to be replaced.

Limitations that are not addressed as part of this are:
- A "king" piece type with a constant count of 1 is required.
- Pockets are not considered.