Update variant.cpp
Added Musketeer Chess Variant description. The Gating is like Seirawan Chess but the squares for Gating are Selected before the game starts. The default piece combination is Leopard and Musketeer Cannon. This piece combination can be changed in variants.ini 
When Pychess will be deployed, a trick should be found to allow piece selection from a bunch of possible choices.